// packages/core/src/chat.d.ts
export type EmbeddingFunction = (input: string | string[], cfg: LanguageModelConfiguration, options: TraceOptions & CancellationOptions & RetryOptions) => Promise<EmbeddingResult>;
export type WorkspaceFileIndexCreator = (indexName: string, cfg: LanguageModelConfiguration, embedder: EmbeddingFunction, options?: VectorIndexOptions & TraceOptions & CancellationOptions) => Promise<WorkspaceFileIndex>;